+++ RaylibStarterCPP1/SonicChaos/EnemyAgent.cpp
@@ -52,5 +52,5 @@ void EnemyAgent::Update(float deltaTime)
 //---------------------
 void EnemyAgent::Draw()
 {
-	DrawTexture(m_texture, m_position.x, m_position.y, WHITE);
+	DrawTexture(m_texture, (int)m_position.x, (int)m_position.y, WHITE);
 }
