+++ RaylibStarterCPP1/SonicChaos/PlayerAgent.cpp
@@ -52,5 +52,5 @@ void PlayerAgent::Update(float deltaTime)
 //----------------------
 void PlayerAgent::Draw()
 {
-	DrawTexture(m_texture, m_position.x, m_position.y, WHITE);
+	DrawTexture(m_texture, (int)m_position.x, (int)m_position.y, WHITE);
 }
